match_delegate: move to extensions
