+++ CODEOWNERS
@@ -339,6 +339,8 @@ extensions/filters/http/oauth2 @derekargueta @mattklein123
 /*/extensions/geoip_providers/common @nezdolik @ravenblackx
 # Maxmind geolocation provider
 /*/extensions/geoip_providers/maxmind @nezdolik @ravenblackx
+# Match delegate extension
+/*/extensions/filters/http/match_delegate @wbpcode @jstraceski
 
 /*/extensions/health_checkers/common @zuercher @botengyao
