+++ source/exe/BUILD
@@ -162,6 +162,8 @@ envoy_cc_library(
         "//source/extensions/listener_managers/validation_listener_manager:validation_listener_manager_lib",
         "//source/common/version:version_linkstamp",
         "//source/server:options_lib",
+        # TODO(wbpcode) make this a proper extension
+        "//source/extensions/filters/http/match_delegate:config",
     ] + envoy_select_enable_http3(["//source/common/quic:server_codec_lib"]),
 )
