+++ source/extensions/filters/http/match_delegate/BUILD
@@ -4,15 +4,15 @@ load(
 )
 load(
     "//bazel:envoy_build_system.bzl",
-    "envoy_cc_library",
-    "envoy_package",
+    "envoy_cc_extension",
+    "envoy_extension_package",
 )
 
 licenses(["notice"])  # Apache 2
 
-envoy_package()
+envoy_extension_package()
 
-envoy_cc_library(
+envoy_cc_extension(
     name = "config",
     srcs = ["config.cc"],
     hdrs = ["config.h"],
