+++ test/extensions/filters/http/match_delegate/BUILD
@@ -12,7 +12,7 @@ envoy_cc_test(
     name = "config_test",
     srcs = ["config_test.cc"],
     deps = [
-        "//source/common/http/match_delegate:config",
+        "//source/extensions/filters/http/match_delegate:config",
         "//test/mocks/server:factory_context_mocks",
         "//test/test_common:registry_lib",
     ],
@@ -23,6 +23,7 @@ envoy_cc_test(
     srcs = ["match_delegate_integration_test.cc"],
     deps = [
         "//source/extensions/filters/http/common:factory_base_lib",
+        "//source/extensions/filters/http/match_delegate:config",
         "//source/server/config_validation:server_lib",
         "//test/integration:http_integration_lib",
         "//test/integration/filters:set_response_code_filter_config_proto_cc_proto",
